Arabian Dances: Add score outline
The longest piece we listened to in Wind Ensemble 2 searching for our wind festival pieces. Actually we didn't listen all the way through it because some horn player piped up that he played this in eighth grade. And Sybilski is unsure if we want to play yet another Arabic/Middle-Eastern piece after an entire season of marching one.

This one was an annoying one. Forcing me to bring out rhythm_mark.ly from On With the Snow!. It's only used for the 4/2 part and I'm not sure why Balmages made it 4/2. Ah well.