Russian Christmas Music: Add score outline
The long-vaulted piece that is one of the pieces always played during PHS's winter concert! Except when we switched to trimesters they made it so only the class of 2020 and before could play it if someone wasn't in Wind 1. That sucks.

This is still in development so this might be changed a little bit later.