Russian Christmas Music: Add Flute 1
I've been working on this along with the score outline for a while, I'd rather have this get put into a commit and pushed to the development branch now rather than later.

There's still a few issues that need to be fixed, of which I know beaming is an important one.