An Unending Legacy: Add score outline
This was the piece that Sybilski ended up going with for our third lyrical piece for our festival in (or near) April. The other two candidates were "Autumn On White Lake" and "Adagio", but Sybilski said he was already kind of leaning towards "An Unending Legacy" and I was too lazy to engrave the two others so I just did this one.