Variations on a Korean Folk Song: Add score outline
The other piece we're planning on playing for the Poway Festival in Tri 3, along with Alfred Reed's Greensleeves. We were actually handed out the music for this before the February break, but not for Greensleeves.

I haven't actually used this in a part yet, so it could have errors, and I haven't checked it for miscopies. But I found an engraving error in my part: the rehearsal mark at measure 241 should actually say 240 instead.