Variations on a Korean Folk Song: Add note value changes to m. 183 & 199
This is because when I was originally creating the score outline, I didn't know what exactly to do to get this notation in. But I decided to use \mark since there were no rehearsal marks here and it seemed I could create a dotted note with \note anyway.

Note that I still haven't actually tested this yet to see if it works or not.